langchain_transformers/requirements.txt
transformers>=4.40.0
langchain-huggingface>=0.0.21
torch>=2.3.0
pyyaml>=6.0.1
argparse>=1.4.0
tqdm>=4.66.2  # For potential future progress bars